Reversed CW and CCW in code. Seems something is backwards
Wrote docking state machine. Under TEST

+++ RobotNoASF/Interfaces/timer0.c
@@ -149,7 +149,7 @@ int delay_ms(uint32_t period_ms)
 
 /*
 * Function:
-* int fdelay_ms(uint32_t period_ms)
+* uint8_t fdelay_ms(uint32_t period_ms)
 *
 * Multi-task friendly delay
 *
@@ -168,7 +168,7 @@ int delay_ms(uint32_t period_ms)
 * [Ideas for improvements that are yet to be made](optional)
 *
 */
-int fdelay_ms(uint32_t period_ms)
+uint8_t fdelay_ms(uint32_t period_ms)
 {
 	enum {START, WAIT, STOP};
 	static uint8_t delayState = START;
